+++ drivers/gpu/msm/kgsl_cffdump.h
@@ -31,20 +31,20 @@ void kgsl_cffdump_init(void);
 void kgsl_cffdump_destroy(void);
 void kgsl_cffdump_open(struct kgsl_device *device);
 void kgsl_cffdump_close(struct kgsl_device *device);
-void kgsl_cffdump_memcpy(struct kgsl_device *device, unsigned int gpuaddr,
-		unsigned int *ptr, size_t sizebytes);
+void kgsl_cffdump_memcpy(struct kgsl_device *device, uint64_t gpuaddr,
+		unsigned int *ptr, uint64_t sizebytes);
 void kgsl_cffdump_syncmem(struct kgsl_device *,
 	struct kgsl_memdesc *memdesc, uint physaddr, size_t sizebytes,
 	bool clean_cache);
-void kgsl_cffdump_memset(struct kgsl_device *device, uint addr,
+void kgsl_cffdump_memset(struct kgsl_device *device, uint64_t addr,
 			unsigned char value, size_t sizebytes);
 void kgsl_cffdump_regwrite(struct kgsl_device *device, uint addr,
 	uint value);
 void kgsl_cffdump_regpoll(struct kgsl_device *device, uint addr,
 	uint value, uint mask);
 bool kgsl_cffdump_parse_ibs(struct kgsl_device_private *dev_priv,
-	const struct kgsl_memdesc *memdesc, uint gpuaddr, int sizedwords,
-	bool check_only);
+	const struct kgsl_memdesc *memdesc, uint64_t gpuaddr,
+	uint64_t sizedwords, bool check_only);
 void kgsl_cffdump_user_event(struct kgsl_device *device,
 		unsigned int cff_opcode, unsigned int op1,
 		unsigned int op2, unsigned int op3,
@@ -65,7 +65,7 @@ void kgsl_cffdump_printline(int id, uint opcode, uint op1, uint op2,
 	uint op3, uint op4, uint op5);
 
 static inline void kgsl_cffdump_write(struct kgsl_device *device,
-		unsigned int gpuaddr, unsigned int value)
+		uint64_t gpuaddr, unsigned int value)
 {
 	if (!device || !device->cff_dump_enable)
 		return;
@@ -96,13 +96,13 @@ static inline void kgsl_cffdump_close(struct kgsl_device *device)
 }
 
 static inline void kgsl_cffdump_write(struct kgsl_device *device,
-		unsigned int gpuaddr, unsigned int value)
+		uint64_t gpuaddr, unsigned int value)
 {
 	return;
 }
 
 static inline void kgsl_cffdump_memcpy(struct kgsl_device *device,
-		unsigned int gupaddr, unsigned int *ptr, size_t sizebytes)
+		uint64_t gupaddr, unsigned int *ptr, uint64_t sizebytes)
 {
 	return;
 }
@@ -114,8 +114,8 @@ static inline void kgsl_cffdump_syncmem(struct kgsl_device *device,
 	return;
 }
 
-static inline void kgsl_cffdump_memset(struct kgsl_device *device, uint addr,
-		unsigned char ch, size_t sizebytes)
+static inline void kgsl_cffdump_memset(struct kgsl_device *device,
+		uint64_t addr, unsigned char ch, size_t sizebytes)
 {
 	return;
 }
@@ -133,8 +133,8 @@ static inline void kgsl_cffdump_regpoll(struct kgsl_device *device, uint addr,
 }
 
 static inline bool kgsl_cffdump_parse_ibs(struct kgsl_device_private *dev_priv,
-	const struct kgsl_memdesc *memdesc, uint gpuaddr, int sizedwords,
-	bool check_only)
+	const struct kgsl_memdesc *memdesc, uint64_t gpuaddr,
+	uint64_t sizedwords, bool check_only)
 {
 	return false;
 }
